pyproject.toml
[project]
name = "markdown-web-browser"
version = "0.1.0"
description = "Render any URL to Markdown via tiled screenshots + olmOCR"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "granian>=2.1.2",
  "playwright>=1.48",
  "pillow>=11.0",
  "pyvips>=2.2",
  "python-decouple>=3.8",
  "pydantic>=2.8",
  "jinja2>=3.1",
  "httpx>=0.27",
  "sqlmodel>=0.0.22",
  "sqlite-vec>=0.1.1",
  "mistune>=3.0",
  "prometheus-client>=0.20",
  "structlog>=24.1",
]

[project.optional-dependencies]
local-ocr = [
  "vllm>=0.6.0",
  "sglang>=0.3.0",
  "olmocr>=0.4.0",
]

observability = [
  "opentelemetry-sdk>=1.28",
  "opentelemetry-exporter-otlp>=1.28",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pyoxipng>=9.0",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]
